Workaround for when mezzanine is hosted under a different urlspace. (Only tested with Django admin, not grappelli)
The keywords.js file needs to reference a mezzanine URL, from the admin site.
It used to use a hard-coded string with a root-absolute path, but this failed when mezzanine was hosted elsewhere.
Instead, we now reference a global Mezzanine JS object, which is set by template, using URL reversal to find the correct url.
This requires a reworking in how the PageAdmin object accesses its Media settings - previously the list of js files was calculated at module load time, but at this stage the url reversal mechanism won't work, because the urls.py aren't all loaded yet, Instead, we hide the list generation inside a lazy iterator object and create the list on demand.