Put block tags into all blog templates so they can be overridden.  Closes #443.
This resulted in the splitting of the editable field for the title and
publication date in the list page into two editable fields, so they
could be two separate blocks.  I notice that the blog detail page
doesn't have an editable field for the publish date at all, which I
shall address separately.

block tags are namespaced by blog_post_detail_ and blog_post_list_
respectively, and inside the list page, the blocks related to an
individual post are namespaced with blog_post_list_post_.