Update mezzanine/blog/models.py
wrong keyword argument passed to blog_post_list_category in get_absolute_url for the BlogCategory model. This results in an empty url when using <a hreg="{{ category.get_absolute_url }}">Link to my Category</a>
The problem was that the blog/urls.py uses 'category' as the keyword and the get_absolute_url used 'slug' as the keyword. I changed it within get_absolute_url because I guess changing it within blog/urls.py may break backwards compatibility.