Add support for settings.RICHTEXT_FILTERS
RICHTEXT_FILTERS is a list of items that are valid for the
RICHTEXT_FILTER setting.  The plural version takes precedence if it is
available and non-empty.  Each item in RICHTEXT_FILTERS is applied in
order.

An alias for the ``richtext_filter`` template filter has been added to
match the plural nature of the new setting: ``richtext_filters`` simply
calls on ``richtext_filter`` for its output.